+++ TinniTrack/Features/Dashboard/Views/StudyTaskDashboardView.swift
@@ -1,18 +1,6 @@
-//
-//  StudyTaskDashboardView.swift
-//  TinniTrack
-//
-
 import SwiftUI
-import UIKit
 
 struct StudyTaskDashboardView: View {
-    private enum OrientationStep {
-        case hearingTest
-        case importAudiogram
-        case nextSteps
-    }
-
     let study: Study
     let enrollment: StudyEnrollment
     let profileTimezone: String?
@@ -20,7 +8,7 @@ struct StudyTaskDashboardView: View {
     @Environment(\.openURL) private var openURL
     @StateObject private var viewModel: StudyTaskDashboardViewModel
     @State private var isOrientationPresented = false
-    @State private var orientationStep: OrientationStep = .hearingTest
+    @State private var orientationStep: StudyTaskOrientationStep = .hearingTest
     @State private var selectedTask: ScheduledTask?
 
     private let studyService: StudyServiceProtocol
@@ -32,16 +20,17 @@ struct StudyTaskDashboardView: View {
         coordinator: AudiogramImportCoordinating = AudiogramImportCoordinator(),
         studyService: StudyServiceProtocol? = nil
     ) {
+        let resolvedStudyService = studyService ?? SupabaseStudyService()
         self.study = study
         self.enrollment = enrollment
         self.profileTimezone = profileTimezone
-        self.studyService = studyService ?? SupabaseStudyService()
+        self.studyService = resolvedStudyService
         _viewModel = StateObject(
             wrappedValue: StudyTaskDashboardViewModel(
                 study: study,
                 enrollment: enrollment,
                 coordinator: coordinator,
-                studyService: studyService ?? SupabaseStudyService(),
+                studyService: resolvedStudyService,
                 profileTimezone: profileTimezone
             )
         )
@@ -66,14 +55,17 @@ struct StudyTaskDashboardView: View {
                     Task { await viewModel.didSubmitTask() }
                 }
             }
-            .alert("Unable to Continue", isPresented: Binding(
-                get: { viewModel.taskLoadErrorMessage != nil },
-                set: { shouldShow in
-                    if !shouldShow {
-                        viewModel.dismissTaskError()
+            .alert(
+                "Unable to Continue",
+                isPresented: Binding(
+                    get: { viewModel.taskLoadErrorMessage != nil },
+                    set: { shouldShow in
+                        if !shouldShow {
+                            viewModel.dismissTaskError()
+                        }
                     }
-                }
-            )) {
+                )
+            ) {
                 Button("OK", role: .cancel) {
                     viewModel.dismissTaskError()
                 }
@@ -89,7 +81,7 @@ struct StudyTaskDashboardView: View {
         } else {
             switch viewModel.contentState {
             case .loading:
-                ProgressView("Checking study prerequisites…")
+                ProgressView("Checking study prerequisites...")
                     .frame(maxWidth: .infinity, maxHeight: .infinity)
                     .background(Color(uiColor: .systemGroupedBackground))
             case .blocked:
@@ -103,12 +95,12 @@ struct StudyTaskDashboardView: View {
     private var orientationRequiredContent: some View {
         ScrollView {
             VStack(alignment: .leading, spacing: 18) {
-                prerequisiteCard(
+                StudyPrerequisiteCard(
                     title: "Welcome. Thanks for choosing to participate in this study!",
                     message: "Before tasks can start, complete orientation and import your hearing-test baseline."
                 )
 
-                actionButton(title: "Begin Orientation", isPrimary: true) {
+                StudyActionButton(title: "Begin Orientation", isPrimary: true) {
                     orientationStep = .hearingTest
                     isOrientationPresented = true
                 }
@@ -121,12 +113,12 @@ struct StudyTaskDashboardView: View {
     private var blockedPrerequisiteContent: some View {
         ScrollView {
             VStack(alignment: .leading, spacing: 18) {
-                prerequisiteCard(
+                StudyPrerequisiteCard(
                     title: "Study Tasks Are Temporarily Locked",
                     message: "A hearing-test baseline is required before you can run loudness tasks."
                 )
 
-                actionButton(title: "Resolve Prerequisites", isPrimary: true) {
+                StudyActionButton(title: "Resolve Prerequisites", isPrimary: true) {
                     orientationStep = .importAudiogram
                     isOrientationPresented = true
                 }
@@ -140,7 +132,7 @@ struct StudyTaskDashboardView: View {
         List {
             if let warning = viewModel.readySyncWarning {
                 Section {
-                    readyWarningCard(for: warning)
+                    ReadySyncWarningCard(warning: warning)
                 }
             }
 
@@ -160,7 +152,7 @@ struct StudyTaskDashboardView: View {
                             if viewModel.isSyncing {
                                 ProgressView()
                             }
-                            Text(viewModel.isSyncing ? "Syncing from Health…" : "Sync from Health")
+                            Text(viewModel.isSyncing ? "Syncing from Health..." : "Sync from Health")
                         }
                     }
                     .disabled(viewModel.isSyncing)
@@ -171,7 +163,7 @@ struct StudyTaskDashboardView: View {
                 if viewModel.isLoadingTasks {
                     HStack {
                         ProgressView()
-                        Text("Loading tasks…")
+                        Text("Loading tasks...")
                             .foregroundStyle(.secondary)
                     }
                 } else if viewModel.futureTasks.isEmpty {
@@ -179,7 +171,11 @@ struct StudyTaskDashboardView: View {
                         .foregroundStyle(.secondary)
                 } else {
                     ForEach(viewModel.futureTasks) { task in
-                        futureTaskRow(task)
+                        FutureStudyTaskRow(
+                            task: task,
+                            canStart: viewModel.canStart(task),
+                            onStart: { selectedTask = task }
+                        )
                     }
                 }
             }
@@ -190,7 +186,7 @@ struct StudyTaskDashboardView: View {
                         .foregroundStyle(.secondary)
                 } else {
                     ForEach(viewModel.completedTasks) { task in
-                        completedTaskRow(task)
+                        CompletedStudyTaskRow(task: task)
                     }
                 }
             }
@@ -198,361 +194,18 @@ struct StudyTaskDashboardView: View {
         .listStyle(.insetGrouped)
     }
 
-    private func futureTaskRow(_ task: ScheduledTask) -> some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text("Day \(task.dayIndex + 1), Slot \(task.slotIndex + 1)")
-                .font(.subheadline)
-                .fontWeight(.semibold)
-
-            Text(Self.windowFormatter.string(from: task.windowStart) + " - " + Self.timeFormatter.string(from: task.windowEnd))
-                .font(.footnote)
-                .foregroundStyle(.secondary)
-
-            Button {
-                selectedTask = task
-            } label: {
-                Text("Start Task")
-                    .frame(maxWidth: .infinity)
-                    .padding(.vertical, 10)
-            }
-            .buttonStyle(.borderedProminent)
-            .disabled(!viewModel.canStart(task))
-
-            if !viewModel.canStart(task) {
-                Text(startAvailabilityMessage(for: task))
-                    .font(.footnote)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .padding(.vertical, 4)
-    }
-
-    private func completedTaskRow(_ task: ScheduledTask) -> some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text("Day \(task.dayIndex + 1), Slot \(task.slotIndex + 1)")
-                .font(.subheadline)
-                .fontWeight(.semibold)
-
-            if let completedAt = task.completedAt {
-                Text("Completed \(Self.dateFormatter.string(from: completedAt))")
-                    .font(.footnote)
-                    .foregroundStyle(.secondary)
-            } else {
-                Text("Completed")
-                    .font(.footnote)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .padding(.vertical, 2)
-    }
-
     private var orientationSheet: some View {
-        NavigationStack {
-            ScrollView {
-                VStack(alignment: .leading, spacing: 18) {
-                    switch orientationStep {
-                    case .hearingTest:
-                        orientationHearingTestStep
-                    case .importAudiogram:
-                        orientationImportStep
-                    case .nextSteps:
-                        orientationNextSteps
-                    }
-                }
-                .padding(20)
-            }
-            .background(Color(uiColor: .systemGroupedBackground))
-            .navigationTitle("Study Orientation")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                if orientationStep != .hearingTest {
-                    ToolbarItem(placement: .topBarLeading) {
-                        Button {
-                            moveToPreviousOrientationStep()
-                        } label: {
-                            Image(systemName: "chevron.left")
-                                .font(.headline)
-                        }
-                        .accessibilityLabel("Back")
-                    }
-                }
-                ToolbarItem(placement: .topBarTrailing) {
-                    Button("Close") {
-                        isOrientationPresented = false
-                    }
-                }
-            }
-        }
-        .interactiveDismissDisabled(true)
-        .presentationDetents([.large])
-        .presentationDragIndicator(.hidden)
+        StudyTaskOrientationSheet(
+            step: $orientationStep,
+            viewModel: viewModel,
+            openHealthApp: openHealthApp,
+            close: { isOrientationPresented = false }
+        )
     }
 
-    private var orientationHearingTestStep: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Step 1")
-                .font(.caption)
-                .fontWeight(.semibold)
-                .tracking(0.8)
-                .foregroundStyle(.secondary)
-
-            hearingTestInstructions
-
-            actionButton(title: "Continue", isPrimary: true) {
-                orientationStep = .importAudiogram
-            }
-        }
-    }
-
-    private var orientationImportStep: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Step 2")
-                .font(.caption)
-                .fontWeight(.semibold)
-                .tracking(0.8)
-                .foregroundStyle(.secondary)
-
-            importStateContent
-
-            actionButton(title: "Continue", isPrimary: true) {
-                orientationStep = .nextSteps
-            }
-            .disabled(!viewModel.isAudiogramPrerequisiteMet)
-            .opacity(viewModel.isAudiogramPrerequisiteMet ? 1 : 0.5)
-        }
-        .task {
-            await viewModel.checkOrientationImportStatus()
-        }
-    }
-
-    private var orientationNextSteps: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Step 3")
-                .font(.caption)
-                .fontWeight(.semibold)
-                .tracking(0.8)
-                .foregroundStyle(.secondary)
-
-            prerequisiteCard(
-                title: "Finish Orientation",
-                message: "Once you finish, we will generate your full Study No. 1 task schedule."
-            )
-
-            actionButton(
-                title: viewModel.isCompletingStudyOnboarding ? "Finishing…" : "Finish Orientation",
-                isPrimary: true,
-                isLoading: viewModel.isCompletingStudyOnboarding
-            ) {
-                Task {
-                    await viewModel.completeStudyOnboarding()
-                    if !viewModel.requiresStudyOnboardingCompletion {
-                        isOrientationPresented = false
-                    }
-                }
-            }
-            .disabled(!viewModel.isAudiogramPrerequisiteMet || viewModel.isCompletingStudyOnboarding)
-            .opacity((viewModel.isAudiogramPrerequisiteMet && !viewModel.isCompletingStudyOnboarding) ? 1 : 0.5)
-        }
-    }
-
-    @ViewBuilder
-    private var importStateContent: some View {
-        switch viewModel.orientationImportState {
-        case .waitingForPermission:
-            prerequisiteCard(
-                title: "Connect Apple Health",
-                message: "Allow Apple Health access so we can import your hearing test into the study."
-            )
-
-            actionButton(
-                title: viewModel.isSyncing ? "Connecting…" : "Connect Apple Health",
-                isPrimary: true,
-                isLoading: viewModel.isSyncing
-            ) {
-                Task { await viewModel.connectAppleHealthForOrientation() }
-            }
-        case .requestingOrChecking:
-            ProgressView("Checking hearing test import…")
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-        case .success(let hearingTestDate):
-            prerequisiteCard(
-                title: "Hearing Test Imported",
-                message: successMessageForHearingTestDate(hearingTestDate)
-            )
-
-        case .authorizedNoHearingTest:
-            prerequisiteCard(
-                title: "No Hearing Test Found Yet",
-                message: "We can access your health data, but we are not seeing a hearing test yet."
-            )
-
-            hearingTestInstructions
-
-            actionButton(
-                title: viewModel.isSyncing ? "Checking…" : "Check Again",
-                isPrimary: true,
-                isLoading: viewModel.isSyncing
-            ) {
-                Task { await viewModel.connectAppleHealthForOrientation() }
-            }
-
-        case .permissionDenied:
-            prerequisiteCard(
-                title: "Permission Required",
-                message: "Approve hearing-test access in Apple Health, then return here and check again."
-            )
-
-            actionButton(title: "Open Health App", isPrimary: false) {
-                guard let healthAppURL = URL(string: "x-apple-health://") else {
-                    return
-                }
-                openURL(healthAppURL)
-            }
-
-            actionButton(
-                title: viewModel.isSyncing ? "Checking…" : "Check Again",
-                isPrimary: true,
-                isLoading: viewModel.isSyncing
-            ) {
-                Task { await viewModel.checkOrientationImportStatus() }
-            }
-
-        case .error(let message):
-            prerequisiteCard(
-                title: "Import Unavailable",
-                message: message
-            )
-
-            actionButton(
-                title: viewModel.isSyncing ? "Retrying…" : "Try Again",
-                isPrimary: true,
-                isLoading: viewModel.isSyncing
-            ) {
-                Task { await viewModel.connectAppleHealthForOrientation() }
-            }
-        }
-    }
-
-    private var hearingTestInstructions: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            prerequisiteCard(
-                title: "Take an Apple Hearing Test",
-                message: "Use AirPods Pro with your paired iPhone. In Settings > your AirPods, tap Take a Hearing Test. Then come back and continue orientation."
-            )
-
-            Link(
-                "Need help taking an Apple Hearing Test?",
-                destination: URL(string: "https://support.apple.com/en-us/120991")!
-            )
-            .font(.subheadline)
-            .fontWeight(.semibold)
-        }
-    }
-
-    private func prerequisiteCard(title: String, message: String) -> some View {
-        VStack(alignment: .leading, spacing: 10) {
-            Text(title)
-                .font(.headline)
-            Text(message)
-                .font(.subheadline)
-                .foregroundStyle(.secondary)
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(Color.white)
-        .clipShape(RoundedRectangle(cornerRadius: 14, style: .continuous))
-        .shadow(color: .black.opacity(0.08), radius: 3, x: 0, y: 1)
-    }
-
-    private func actionButton(
-        title: String,
-        isPrimary: Bool,
-        isLoading: Bool = false,
-        action: @escaping () -> Void
-    ) -> some View {
-        Button(action: action) {
-            HStack {
-                if isLoading {
-                    ProgressView()
-                        .tint(isPrimary ? .white : .blue)
-                }
-                Text(title)
-                    .fontWeight(.semibold)
-            }
-            .frame(maxWidth: .infinity)
-            .padding(.vertical, 14)
-        }
-        .buttonStyle(.plain)
-        .background(isPrimary ? Color.blue : Color.white)
-        .foregroundStyle(isPrimary ? Color.white : Color.blue)
-        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-        .shadow(color: .black.opacity(isPrimary ? 0 : 0.08), radius: 3, x: 0, y: 1)
-    }
-
-    private func successMessageForHearingTestDate(_ date: Date?) -> String {
-        guard let date else {
-            return "Success! We got your hearing test."
-        }
-        return "Success! We got your hearing test from \(Self.hearingTestDateFormatter.string(from: date))."
-    }
-
-    private func readyWarningCard(
-        for warning: StudyTaskDashboardViewModel.ReadySyncWarning
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(alignment: .firstTextBaseline, spacing: 8) {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundStyle(.orange)
-                Text(readyWarningTitle(for: warning))
-                    .font(.subheadline)
-                    .fontWeight(.semibold)
-            }
-
-            Text(readyWarningMessage(for: warning))
-                .font(.footnote)
-                .foregroundStyle(.secondary)
-        }
-        .padding(.vertical, 8)
-    }
-
-    private func readyWarningTitle(
-        for warning: StudyTaskDashboardViewModel.ReadySyncWarning
-    ) -> String {
-        switch warning {
-        case .permissionDenied:
-            return "Health Access Needed for Sync"
-        case .noAudiogramInHealth:
-            return "No New Hearing Test Found"
-        case .error:
-            return "Unable to Sync from Health"
-        }
-    }
-
-    private func readyWarningMessage(
-        for warning: StudyTaskDashboardViewModel.ReadySyncWarning
-    ) -> String {
-        switch warning {
-        case .permissionDenied:
-            return "Tasks remain available, but we could not read hearing-test data. Re-enable Health access and tap Sync from Health again."
-        case .noAudiogramInHealth:
-            return "We can access your health data, but we are not seeing a hearing test yet."
-        case .error(let message):
-            return message
-        }
-    }
-
-    private func startAvailabilityMessage(for task: ScheduledTask) -> String {
-        let now = Date()
-        if now < task.windowStart {
-            return "Available at \(Self.windowFormatter.string(from: task.windowStart))."
-        }
-
-        if now > task.windowEnd {
-            return "This task window has ended."
-        }
-
-        return "This task is temporarily unavailable."
+    private func openHealthApp() {
+        guard let healthAppURL = URL(string: "x-apple-health://") else { return }
+        openURL(healthAppURL)
     }
 
     private func handleOrientationDismissed() {
@@ -560,45 +213,10 @@ struct StudyTaskDashboardView: View {
         Task { await viewModel.refresh() }
     }
 
-    private func moveToPreviousOrientationStep() {
-        switch orientationStep {
-        case .hearingTest:
-            break
-        case .importAudiogram:
-            orientationStep = .hearingTest
-        case .nextSteps:
-            orientationStep = .importAudiogram
-        }
-    }
-
     private static let dateFormatter: DateFormatter = {
         let formatter = DateFormatter()
         formatter.locale = Locale(identifier: "en_US_POSIX")
         formatter.dateStyle = .medium
-        formatter.timeStyle = .short
-        return formatter
-    }()
-
-    private static let hearingTestDateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.dateStyle = .long
-        formatter.timeStyle = .none
-        return formatter
-    }()
-
-    private static let windowFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.dateStyle = .medium
-        formatter.timeStyle = .short
-        return formatter
-    }()
-
-    private static let timeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.dateStyle = .none
         formatter.timeStyle = .short
         return formatter
     }()
